+++ src/models/spriteMarker.ts
@@ -1,0 +1,9 @@
+import * as PIXI from 'pixi.js'
+
+export interface SpriteMarker {
+  id: number
+  sprite: PIXI.Sprite
+  popup: L.Popup
+  position: L.LatLngTuple
+  size: number
+}
